Finish debug

+++ include/Algorithm/DataStructure/MicroCluster.hpp
@@ -30,7 +30,7 @@ class MicroCluster {
   dataPoint SS;// the sum of the squares of  the data values for each dimension
   int LST; //  the sum of the time stamps Ti~... Tin
   int SST;//the sum of the squares of the time stamps  Til... Tin
-  int weight; //number of data point in the clusters
+  double weight; //number of data point in the clusters
   int dimension;
 
   //TODO Need to subtract Base class of CF vector when all cf-vector based-algorithms have been implemented
